% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_for_status.R
\name{check_for_status}
\alias{check_for_status}
\title{check_for_status}
\usage{
check_for_status(x, type = "error", message_function = message_standard,
  ...)
}
\arguments{
\item{type}{character, one of "error", "warning", "message"}

\item{message_function}{function, builds message,
takes \code{x} as first arguement, \code{...} as additional arguements.
Default is a standard message function.}

\item{...}{additional arguements for \code{message_function}}
}
\value{
If request was successful, an invisible TRUE.
   Otherwise, raised a classed http error or warning, as generated by \code{http_condition}
}
\description{
check_for_status
}
\examples{
x <- httr::GET("http://httpbin.org/status/404")
  check_for_status(x, type = "error", message_function = message_verbose)
}

