% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/message.R
\name{message_standard}
\alias{message_standard}
\title{message_standard}
\usage{
message_standard(x)
}
\arguments{
\item{x}{a request object}
}
\value{
character, single string concerning the request, suitable input to \code{cat()}
}
\description{
This is a recreation of the message used in the \code{httr} package.
}
\examples{
x <- httr::GET("http://httpbin.org/status/200")
  message_standard(x)
}

