% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/message.R
\name{message_verbose}
\alias{message_verbose}
\title{message_verbose}
\usage{
message_verbose(x)
}
\arguments{
\item{x}{a request object}
}
\value{
single string concerning the request, suitable input to \code{cat()}
}
\description{
A more-verbose message.
}
\details{
This returns a more verbose message that includes:

\itemize{
  \item standard message
  \item date
  \item url
}
}
\examples{
x <- httr::GET("http://httpbin.org/status/200")
  message_verbose(x)
}

